+++ NEWS.md
@@ -5,6 +5,7 @@
 - Use unicode symbols only on terminals that support UTF8, and respect the
   `cli.unicode` option, so that `options(cli.unicode = FALSE)` can be used to
   disable all unicode output.
+- Ignore deprecated functions by default in `get_exported_functions()`.
 
 # CBTF 0.5.0 (2025-08-21)